Refactoring (#197)
* atmos module and indices refactor

* Indices refactoring and small changes

* Doc fixes

* Reserved namespaces and ignoring flake8 errors
