+++ docs/indices.rst
@@ -2,9 +2,20 @@ All Indices
 ===========
 
 
-.. automodule:: xclim.indices
+.. automodule:: xclim.indices._simple
    :members:
    :undoc-members:
    :show-inheritance:
    :noindex:
 
+.. automodule:: xclim.indices._multivariate
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :noindex:
+
+.. automodule:: xclim.indices._threshold
+   :members:
+   :undoc-members:
+   :show-inheritance:
+   :noindex:
