+++ setup.cfg
@@ -33,6 +33,10 @@ exclude =
 max-line-length = 120
 max-complexity = 10
 
+ignore =
+    F401
+    F403
+
 [aliases]
 test = pytest
