+++ tests/test_checks.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 import pytest
 import xarray as xr
-from xclim.temperature import tg_mean
+from xclim.atmos import tg_mean
 from xclim.testing.common import tas_series, tasmin_series
 from xclim import checks
