+++ xclim/atmos/__init__.py
@@ -1,0 +1,14 @@
+"""
+Atmospheric indicator calculation instances
+===========================================
+
+While the `indices` module stores the computing functions, this module defines Indicator classes and instances that
+include a number of functionalities, such as input validation, unit conversion, output meta-data handling,
+and missing value masking.
+
+The concept followed here is to define Indicator subclasses for each input variable, then create instances
+for each indicator.
+
+"""
+from ._temperature import *
+from ._precip import *
