+++ xclim/utils.py
@@ -4,22 +4,22 @@
 xclim xarray.DataArray utilities module
 """
 
-import numpy as np
-import six
-import re
-import pint
-import pandas as pd
-import xarray as xr
-from . import checks
-from inspect import signature
-import functools
 import abc
-from collections import defaultdict
-import datetime as dt
 import calendar
-from pyproj import Geod
+import datetime as dt
+import functools
+import re
 import warnings
+from collections import defaultdict
+from inspect import signature
+
+import numpy as np
+import pint
+import six
+import xarray as xr
 from boltons.funcutils import wraps
+
+from . import checks
 
 units = pint.UnitRegistry(autoconvert_offset_to_baseunit=True)
 units.define(pint.unit.UnitDefinition('percent', '%', (),
@@ -287,275 +287,6 @@ def declare_units(out_units, **units_by_name):
     return dec
 
 
-def create_ensemble(ncfiles, mf_flag=False):
-    """Create an xarray datset of ensemble of climate simulation from a list of netcdf files. Input data is
-    concatenated along a newly created data dimension ('realization')
-
-    Returns a xarray dataset object containing input data from the list of netcdf files concatenated along
-    a new dimension (name:'realization'). In the case where input files have unequal time dimensions output
-    ensemble dataset is created for overlapping time-steps common to all input files
-
-    Parameters
-    ----------
-    ncfiles : sequence
-      List of netcdf file paths. If mf_flag is true ncfiles should be a list of lists where
-    each sublist contains input .nc files of a multifile dataset
-
-    mf_flag : Boolean . If true climate simulations are treated as multifile datasets before concatenation
-
-    Returns
-    -------
-    xarray dataset containing concatenated data from all input files
-
-    Notes
-    -----
-    Input netcdf files require equal spatial dimension size (e.g. lon, lat dimensions)
-    If input data contains multiple cftime calendar types they must be at monthly or coarser frequency
-
-    Examples
-    --------
-    >>> from xclim import utils
-    >>> import glob
-    >>> ncfiles = glob.glob('/*.nc')
-    >>> ens = utils.create_ensemble(ncfiles)
-    >>> print(ens)
-    Using multifile datasets:
-    simulation 1 is a list of .nc files (e.g. separated by time)
-    >>> ncfiles = glob.glob('dir/*.nc')
-    simulation 2 is also a list of .nc files
-    >>> ens = utils.create_ensemble(ncfiles)
-    """
-    dim = 'realization'
-    ds1 = []
-    start_end_flag = True
-    print('finding common time-steps')
-    for n in ncfiles:
-        if mf_flag:
-            ds = xr.open_mfdataset(n, concat_dim='time', decode_times=False, chunks={'time': 10})
-            ds['time'] = xr.open_mfdataset(n).time
-        else:
-            ds = xr.open_dataset(n, decode_times=False)
-            ds['time'] = xr.decode_cf(ds).time
-        # get times - use common
-        time1 = pd.to_datetime({'year': ds.time.dt.year, 'month': ds.time.dt.month, 'day': ds.time.dt.day})
-        if start_end_flag:
-            start1 = time1.values[0]
-            end1 = time1.values[-1]
-            start_end_flag = False
-        if time1.values.min() > start1:
-            start1 = time1.values.min()
-        if time1.values.max() < end1:
-            end1 = time1.values.max()
-
-    for n in ncfiles:
-        print('accessing file ', ncfiles.index(n) + 1, ' of ', len(ncfiles))
-        if mf_flag:
-            ds = xr.open_mfdataset(n, concat_dim='time', decode_times=False, chunks={'time': 10})
-            ds['time'] = xr.open_mfdataset(n).time
-        else:
-            ds = xr.open_dataset(n, decode_times=False, chunks={'time': 10})
-            ds['time'] = xr.decode_cf(ds).time
-
-        ds['time'].values = pd.to_datetime({'year': ds.time.dt.year, 'month': ds.time.dt.month, 'day': ds.time.dt.day})
-
-        ds = ds.where((ds.time >= start1) & (ds.time <= end1), drop=True)
-
-        ds1.append(ds.drop('time'))
-    print('concatenating files : adding dimension ', dim, )
-    ens = xr.concat(ds1, dim=dim)
-    # assign time coords
-    ens = ens.assign_coords(time=ds.time.values)
-    return ens
-
-
-def ensemble_percentiles(ens, values=(10, 50, 90), time_block=None):
-    """Calculate ensemble statistics between a results from an ensemble of climate simulations
-
-    Returns a dataset containing ensemble statistics for input climate simulations.
-    Alternatively calculate ensemble percentiles (default) or ensemble mean and standard deviation
-
-    Parameters
-    ----------
-    ens : Ensemble dataset (see xclim.utils.create_ensemble)
-    values : tuple of integers - percentile values to calculate  : default : (10, 50, 90)
-    time_block : integer
-      for large ensembles iteratively calculate percentiles in time-step blocks (n==time_block).
-       If not defined the function tries to estimate an appropriate value
-
-    Returns
-    -------
-    xarray dataset with containing data variables of requested ensemble statistics
-
-    Examples
-    --------
-    >>> from xclim import utils
-    >>> import glob
-    >>> ncfiles = glob.glob('/*tas*.nc')
-    Create ensemble dataset
-    >>> ens = utils.create_ensemble(ncfiles)
-    Calculate default ensemble percentiles
-    >>> ens_percs = utils.ensemble_statistics(ens)
-    >>> print(ens_percs['tas_p10'])
-    Calculate non-default percentiles (25th and 75th)
-    >>> ens_percs = utils.ensemble_statistics(ens, values=(25,75))
-    >>> print(ens_percs['tas_p25'])
-    Calculate by time blocks (n=10) if ensemble size is too large to load in memory
-    >>> ens_percs = utils.ensemble_statistics(ens, time_block=10)
-    >>> print(ens_percs['tas_p25'])
-
-    """
-
-    ds_out = ens.drop(ens.data_vars)
-    dims = list(ens.dims)
-    for v in ens.data_vars:
-        # Percentile calculation requires load to memory : automate size for large ensemble objects
-        if not time_block:
-            time_block = round(2E8 / (ens[v].size / ens[v].shape[dims.index('time')]), -1)  # 2E8
-
-        if time_block > len(ens[v].time):
-            out = calc_percentiles_simple(ens, v, values)
-
-        else:
-            # loop through blocks
-            Warning('large ensemble size detected : statistics will be calculated in blocks of ', int(time_block),
-                    ' time-steps')
-            out = calc_percentiles_blocks(ens, v, values, time_block)
-        for vv in out.data_vars:
-            ds_out[vv] = out[vv]
-    return ds_out
-
-
-def calc_percentiles_simple(ens, v, values):
-    ds_out = ens.drop(ens.data_vars)
-    dims = list(ens[v].dims)
-    outdims = [x for x in dims if 'realization' not in x]
-
-    print('loading ensemble data to memory')
-    arr = ens[v].load()  # percentile calc requires loading the array
-    coords = {}
-    for c in outdims:
-        coords[c] = arr[c]
-    for p in values:
-        outvar = v + '_p' + str(p)
-
-        out1 = calc_perc(arr, p)
-
-        ds_out[outvar] = xr.DataArray(out1, dims=outdims, coords=coords)
-        ds_out[outvar].attrs = ens[v].attrs
-        if 'description' in ds_out[outvar].attrs.keys():
-            ds_out[outvar].attrs['description'] = ds_out[outvar].attrs['description'] + ' : ' + str(p) + \
-                                                  'th percentile of ensemble'
-        else:
-            ds_out[outvar].attrs['description'] = str(p) + \
-                                                  'th percentile of ensemble'
-    return ds_out
-
-
-def calc_percentiles_blocks(ens, v, values, time_block):
-    ds_out = ens.drop(ens.data_vars)
-    dims = list(ens[v].dims)
-    outdims = [x for x in dims if 'realization' not in x]
-
-    blocks = list(range(0, len(ens.time) + 1, int(time_block)))
-    if blocks[-1] != len(ens[v].time):
-        blocks.append(len(ens[v].time))
-    arr_p_all = {}
-    for t in range(0, len(blocks) - 1):
-        print('Calculating block ', t + 1, ' of ', len(blocks) - 1)
-        time_sel = slice(blocks[t], blocks[t + 1])
-        arr = ens[v].isel(time=time_sel).load()  # percentile calc requires loading the array
-        coords = {}
-        for c in outdims:
-            coords[c] = arr[c]
-        for p in values:
-
-            out1 = calc_perc(arr, p)
-
-            if t == 0:
-                arr_p_all[str(p)] = xr.DataArray(out1, dims=outdims, coords=coords)
-            else:
-                arr_p_all[str(p)] = xr.concat([arr_p_all[str(p)],
-                                               xr.DataArray(out1, dims=outdims, coords=coords)], dim='time')
-    for p in values:
-        outvar = v + '_p' + str(p)
-        ds_out[outvar] = arr_p_all[str(p)]
-        ds_out[outvar].attrs = ens[v].attrs
-        if 'description' in ds_out[outvar].attrs.keys():
-            ds_out[outvar].attrs['description'] = ds_out[outvar].attrs['description'] + ' : ' + str(p) + \
-                                                  'th percentile of ensemble'
-        else:
-            ds_out[outvar].attrs['description'] = str(p) + \
-                                                  'th percentile of ensemble'
-    return ds_out
-
-
-def calc_perc(arr, p):
-    dims = arr.dims
-    # make sure time is the second dimension
-    if dims.index('time') != 1:
-        dims1 = [dims[dims.index('realization')], dims[dims.index('time')]]
-        for d in dims:
-            if d not in dims1:
-                dims1.append(d)
-        arr = arr.transpose(*dims1)
-        dims = dims1
-
-    nan_count = np.isnan(arr).sum(axis=dims.index('realization'))
-    out = np.percentile(arr, p, axis=dims.index('realization'))
-    if np.any((nan_count > 0) & (nan_count < arr.shape[dims.index('realization')])):
-        arr1 = arr.values.reshape(arr.shape[dims.index('realization')],
-                                  int(arr.size / arr.shape[dims.index('realization')]))
-        # only use nanpercentile where we need it (slow performace compared to standard) :
-        nan_index = np.where((nan_count > 0) & (nan_count < arr.shape[dims.index('realization')]))
-        t = np.ravel_multi_index(nan_index, nan_count.shape)
-        out[np.unravel_index(t, nan_count.shape)] = np.nanpercentile(arr1[:, t], p, axis=dims.index('realization'))
-
-    return out
-
-
-def ensemble_mean_std_max_min(ens):
-    """Calculate ensemble statistics between a results from an ensemble of climate simulations
-
-    Returns a dataset containing ensemble mean, standard-deviation,
-    minimum and maximum for input climate simulations.
-
-    Parameters
-    ----------
-    ens : Ensemble dataset (see xclim.utils.create_ensemble)
-
-    Returns
-    -------
-    xarray dataset with containing data variables of ensemble statistics
-
-    Examples
-    --------
-    >>> from xclim import utils
-    >>> import glob
-    >>> ncfiles = glob.glob('/*tas*.nc')
-    Create ensemble dataset
-    >>> ens = utils.create_ensemble(ncfiles)
-    Calculate ensemble statistics
-    >>> ens_means_std = utils.ensemble_mean_std_max_min(ens)
-    >>> print(ens_mean_std['tas_mean'])
-    """
-    dsOut = ens.drop(ens.data_vars)
-    for v in ens.data_vars:
-
-        dsOut[v + '_mean'] = ens[v].mean(dim='realization')
-        dsOut[v + '_stdev'] = ens[v].std(dim='realization')
-        dsOut[v + '_max'] = ens[v].max(dim='realization')
-        dsOut[v + '_min'] = ens[v].min(dim='realization')
-        for vv in dsOut.data_vars:
-            dsOut[vv].attrs = ens[v].attrs
-
-            if 'description' in dsOut[vv].attrs.keys():
-                vv.split()
-                dsOut[vv].attrs['description'] = dsOut[vv].attrs['description'] + ' : ' + vv.split('_')[
-                    -1] + ' of ensemble'
-
-    return dsOut
-
-
 def threshold_count(da, op, thresh, freq):
     """Count number of days above or below threshold.
 
@@ -715,155 +446,6 @@ def adjust_doy_calendar(source, target):
         return source
 
     return _interpolate_doy_calendar(source, doy_max)
-
-
-def subset_bbox(da, lon_bnds=None, lat_bnds=None, start_yr=None, end_yr=None):
-    """Subset a datarray or dataset spatially (and temporally) using a lat lon bounding box and years selection.
-
-    Return a subsetted data array for grid points falling within a spatial bounding box
-    defined by longitude and latitudinal bounds and for years falling within provided year bounds.
-
-    Parameters
-    ----------
-    arr : xarray.DataArray or xarray.Dataset
-      Input data.
-    lon_bnds : list of floats
-      List of maximum and minimum longitudinal bounds. Optional. Defaults to all longitudes in original data-array.
-    lat_bnds :  list of floats
-      List maximum and minimum latitudinal bounds. Optional. Defaults to all latitudes in original data-array.
-    start_yr : int
-      First year of the subset. Defaults to first year of input.
-    end_yr : int
-      Last year of the subset. Defaults to last year of input.
-
-    Returns
-    -------
-    xarray.DataArray or xarray.DataSet
-      subsetted data array or dataset
-
-    Examples
-    --------
-    >>> from xclim import utils
-    >>> ds = xr.open_dataset('pr.day.nc')
-    Subset lat lon and years
-    >>> prSub = utils.subset_bbox(ds.pr, lon_bnds=[-75,-70],lat_bnds=[40,45],start_yr=1990,end_yr=1999)
-    Subset data array lat, lon and single year
-    >>> prSub = utils.subset_bbox(ds.pr, lon_bnds=[-75,-70],lat_bnds=[40,45],start_yr=1990,end_yr=1990)
-    Subset dataarray single year keep entire lon, lat grid
-    >>> prSub = utils.subset_bbox(ds.pr,start_yr=1990,end_yr=1990) # one year only entire grid
-    Subset multiple variables in a single dataset
-    >>> ds = xr.open_mfdataset(['pr.day.nc','tas.day.nc'])
-    >>> dsSub = utils.subset_bbox(ds,lon_bnds=[-75,-70],lat_bnds=[40,45],start_yr=1990,end_yr=1999)
-    """
-
-    if lon_bnds:
-
-        lon_bnds = np.asarray(lon_bnds)
-        if np.all(da.lon > 0) and np.any(lon_bnds < 0):
-            lon_bnds[lon_bnds < 0] += 360
-        da = da.where((da.lon >= lon_bnds.min()) & (da.lon <= lon_bnds.max()), drop=True)
-
-    if lat_bnds:
-        lat_bnds = np.asarray(lat_bnds)
-        da = da.where((da.lat >= lat_bnds.min()) & (da.lat <= lat_bnds.max()), drop=True)
-
-    if start_yr or end_yr:
-        if not start_yr:
-            start_yr = da.time.dt.year.min()
-        if not end_yr:
-            end_yr = da.time.dt.year.max()
-
-        if start_yr > end_yr:
-            raise ValueError("Start date is after end date.")
-
-        year_bnds = np.asarray([start_yr, end_yr])
-        da = da.where((da.time.dt.year >= year_bnds.min()) & (da.time.dt.year <= year_bnds.max()), drop=True)
-
-    return da
-
-
-def subset_gridpoint(da, lon, lat, start_yr=None, end_yr=None):
-    """Extract a nearest gridpoint from datarray based on lat lon coordinate.
-    Time series can optionally be subsetted by year(s)
-
-    Return a subsetted data array (or dataset) for the grid point falling nearest the input
-    longitude and latitudecoordinates. Optionally subset the data array for years falling
-    within provided year bounds
-
-    Parameters
-    ----------
-    da : xarray.DataArray or xarray.DataSet
-      Input data.
-    lon : float
-      Longitude coordinate.
-    lat:  float
-      Latitude coordinate.
-    start_yr : int
-      First year of the subset. Defaults to first year of input.
-    end_yr : int
-      Last year of the subset. Defaults to last year of input.
-
-    Returns
-    -------
-    xarray.DataArray or xarray.DataSet
-      Subsetted data array or dataset
-
-    Examples
-    --------
-    >>> from xclim import utils
-    >>> ds = xr.open_dataset('pr.day.nc')
-    Subset lat lon point and multiple years
-    >>> prSub = utils.subset_gridpoint(ds.pr, lon=-75,lat=45,start_yr=1990,end_yr=1999)
-    Subset lat, lon point and single year
-    >>> prSub = utils.subset_gridpoint(ds.pr, lon=-75,lat=45,start_yr=1990,end_yr=1990)
-     Subset multiple variables in a single dataset
-    >>> ds = xr.open_mfdataset(['pr.day.nc','tas.day.nc'])
-    >>> dsSub = utils.subset_gridpoint(ds, lon=-75,lat=45,start_yr=1990,end_yr=1999)
-    """
-
-    g = Geod(ellps='WGS84')  # WGS84 ellipsoid - decent globaly
-    # adjust for files with all postive longitudes if necessary
-    if np.all(da.lon > 0) and lon < 0:
-        lon += 360
-
-    if len(da.lon.shape) == 1 & len(da.lat.shape) == 1:
-        # create a 2d grid of lon, lat values
-        lon1, lat1 = np.meshgrid(np.asarray(da.lon.values), np.asarray(da.lat.values))
-
-    else:
-        lon1 = da.lon.values
-        lat1 = da.lat.values
-    shp_orig = lon1.shape
-    lon1 = np.reshape(lon1, lon1.size)
-    lat1 = np.reshape(lat1, lat1.size)
-    # calculate geodesic distance between grid points and point of interest
-    az12, az21, dist = g.inv(lon1, lat1, np.broadcast_to(lon, lon1.shape), np.broadcast_to(lat, lat1.shape))
-    dist = dist.reshape(shp_orig)
-
-    iy, ix = np.unravel_index(np.argmin(dist, axis=None), dist.shape)
-    xydims = [x for x in da.dims if 'time' not in x]
-    args = dict()
-    args[xydims[0]] = iy
-    args[xydims[1]] = ix
-    out = da.isel(**args)
-    if start_yr or end_yr:
-        if not start_yr:
-            start_yr = da.time.dt.year.min()
-        if not end_yr:
-            end_yr = da.time.dt.year.max()
-
-        if start_yr > end_yr:
-            raise ValueError("Start date is after end date.")
-
-        year_bnds = np.asarray([start_yr, end_yr])
-
-        if len(year_bnds) == 1:
-            time_cond = da.time.dt.year == year_bnds
-        else:
-            time_cond = (da.time.dt.year >= year_bnds.min()) & (da.time.dt.year <= year_bnds.max())
-        out = out.where(time_cond, drop=True)
-
-    return out
 
 
 def get_daily_events(da, da_value, operator):
